+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to
 ### Changed
 
 - Use anybuf from crates.io
+- drand: Add `QueryMsg::IsAllowlisted` (one word) analogue to
+  `QueryMsg::IsAllowListed`.
 
 ## [0.13.2] - 2023-04-26